use visitor pattern instead of multimethods
The implementation of multimethods I was using has problems with
reload() and packages. So I'm using a *very* basic implementation of
the visitor pattern as I don't want to clobber Score() with methods
for printing humdrum and lily.

Also, the multimethod implementation I was using was covering a bug in
the way Tandem is printed. Ouch!